requirements.txt
Flask
flask-cors
openai
pydub
python-dotenv
elevenlabs